refactor(venue-search): common unified search utils moved to components package
LIIKUNTA-328.
Moved the common utilities of unified search to the components package.
Could not move all the items, because there are some sport specific variables given.
