--- apps/sports-helsinki/src/common/utils/getTranslation.ts
@@ -1,14 +1,0 @@
-import type { Language } from '../../types';
-
-export type TranslationsObject = {
-  fi?: string;
-  en?: string;
-  sv?: string;
-};
-
-export default function getTranslation(
-  translation: TranslationsObject,
-  locale: Language
-) {
-  return translation[locale] ?? translation.fi;
-}
