--- apps/sports-helsinki/src/common/utils/getURLSearchParamsFromAsPath.ts
@@ -1,7 +1,0 @@
-export default function getURLSearchParamsFromAsPath(
-  asPath: string
-): URLSearchParams {
-  const [, searchParams] = asPath.split('?');
-
-  return new URLSearchParams(searchParams);
-}
