+++ apps/sports-helsinki/src/domain/search/venueSearch/SearchPage.tsx
@@ -7,16 +7,16 @@ import {
   SrOnly,
   useIsSmallScreen,
   useSearchTranslation,
-  useUnifiedSearchListQuery,
   useUnifiedSearch,
 } from 'events-helsinki-components';
 import { useRouter } from 'next/router';
 import React from 'react';
 import { useConfig } from 'react-helsinki-headless-cms';
 import { scroller } from 'react-scroll';
-import AppConfig from 'domain/app/AppConfig';
 
 import { removeQueryParamsFromRouter } from 'utils/routerUtils';
+import AppConfig from '../../app/AppConfig';
+import useUnifiedSearchListQuery from '../../unifiedSearch/useUnifiedSearchListQuery';
 import SearchResultsContainer from '../eventSearch/searchResultList/SearchResultsContainer';
 import VenueList from '../venueList/VenueList';
 import styles from './eventSearchPage.module.scss';
