+++ packages/components/src/components/domain/unifiedSearch/index.ts
@@ -1,4 +1,2 @@
 export { default as useUnifiedSearch } from './useUnifiedSearch';
-export * from './useUnifiedSearchVariables';
-export { default as useUnifiedSearchListQuery } from './useUnifiedSearchListQuery';
 export * from './types';
