+++ packages/components/src/utils/__tests__/capitalize.test.ts
