+++ packages/components/src/utils/__tests__/formatToSingleTag.test.ts
