+++ packages/components/src/utils/capitalize.ts
