+++ packages/components/src/utils/formatToSingleTag.ts
