+++ packages/components/src/utils/getTranslation.ts
@@ -1,0 +1,8 @@
+import type { LanguageString } from '../types';
+
+export default function getTranslation(
+  translation: LanguageString,
+  locale: keyof LanguageString
+) {
+  return translation[locale] ?? translation.fi;
+}
